requirements.txt
# Function dependencies, for example:
# package>=version
pyOpenSSL==21.0.0
google-cloud-logging==2.7.0